requirements.txt
streamlit==1.40.1
wordcloud==1.9.4
matplotlib==3.9.2
pillow==11.0.0
numpy==2.1.3
